requirements.txt
Django==1.10.6
djangorestframework==3.6.2
Markdown==2.6.8

